… .claude/settings.local.json
… issues/2025-01-18-natural-language-command-processing.md
… issues/2025-01-19-room-generation-visit-lock-bug.md
… package.json
… src/ai/grokClient.ts
… src/nlp/config.ts
… src/nlp/contextResolver.ts
… src/nlp/enhancedNLPEngine.ts
… src/nlp/localNLPProcessor.ts
… src/nlp/types.ts
… src/nlp/unifiedNLPEngine.ts
… tests/context-resolver-simple.test.ts
… tests/context-resolver.test.ts
… tests/enhanced-nlp-simple.test.ts
… tests/enhanced-nlp.test.ts
… tests/nlp-config.test.ts
… tests/nlp-integration.test.ts
… tests/nlp.test.ts
… tests/unified-nlp.test.ts